컨트롤러 status로 코드 리펙토링
