+++ src/main/java/net/bizLogin/promoter/controller/PmtFacilitatorController.java
@@ -7,12 +7,6 @@ import net.bizLogin.promoter.vo.PmtFacilitatorVO;
 import net.common.common.CommandMap;
 import net.common.common.Status;
 import net.mwav.common.module.Common_Utils;
-import net.mwav.common.module.ValidationLib;
-import org.apache.http.HttpResponse;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpUriRequest;
-import org.apache.http.client.methods.RequestBuilder;
-import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.log4j.Logger;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -20,17 +14,10 @@ import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.annotation.Resource;
-import javax.json.Json;
-import javax.json.JsonObject;
-import javax.json.JsonReader;
-import javax.net.ssl.HttpsURLConnection;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.net.URL;
 import java.util.Map;
 
 
@@ -77,43 +64,43 @@ public class PmtFacilitatorController {
 	@RequestMapping(value = "/bizLogin/promoter/facilitator/pmtFacilitatorLogin.mwav",method = RequestMethod.POST)
 	public ModelAndView selectBizPmtLogin(CommandMap commandMap, HttpServletRequest request, RedirectAttributes rtr) throws Exception {
 		ModelAndView mv = new ModelAndView();
-		boolean chkEmailYN = false;
-
-		// 임시용으로 Controller 로직 삽입
-		final Object token = commandMap.getMap().get("token");
-		if(this.recaptcha(token.toString())){
-			log.info("프로모터 로그인 실패");
-			mv.setViewName("redirect:/Promoter/Facilitator/PmtLogin.mwav");
-			rtr.addFlashAttribute("msg", "로봇으로 감지되었습니다. 다시 시도해주세요");
-		}
 
 		// Promoter 로그인 성공시 값을 가져옴
-		BizPromoter_VO bizPromoterVo = pmtFacilitatorService.selectBizPmtLogin(commandMap.getMap());
+		Map<String, Object> mapVo = pmtFacilitatorService.selectBizPmtLogin(commandMap.getMap());
+		int status = (int) mapVo.get("status");
 
 		// 로그인 성공여부에 따라 페이지및 statusCode, msg 변경
-		if(bizPromoterVo == null){
-			log.info("프로모터 로그인 실패");
-			mv.setViewName("redirect:/Promoter/Facilitator/PmtLogin.mwav");
-			rtr.addFlashAttribute("msg", "비밀번호와 아이디를 확인해주세요");
-		}else {
+		switch(status){
+			case 1:
+				// 로그인한 사용자가 이메일을 인증했는지 검증
+				BizPromoter_VO bizPromoterVo = (BizPromoter_VO) mapVo.get("vo");
+				boolean chkEmailYN = false;
+				chkEmailYN = bizPromoterVo.getPmtCertifyDt() != null ? true : false;
 
-			// 로그인한 사용자가 이메일을 인증했는지 검증
-			chkEmailYN = bizPromoterVo.getPmtCertifyDt() != null ? true : false;
-
-			// 이메일 인증된 사용자의 경우
-			if(chkEmailYN){
-				log.info("프로모터 로그인 성공");
-				mv.setViewName("redirect:/Promoter/Index");
-				request.getSession().setAttribute("bizPromoter", bizPromoterVo);
-			}else{
-				// 이메일 인증이 안되어있는 사용자라면 인증 페이지로 redirect
-				log.info("이메일 인증되지 않은 사용자 로그인");
-				mv.setViewName("redirect:/Promoter/Facilitator/PmtCertifyPage.mwav");
-				rtr.addFlashAttribute("msg", "이메일 인증이 필요합니다.");
-				rtr.addFlashAttribute("promoter_id", bizPromoterVo.getPromoter_id());
-				rtr.addFlashAttribute("pmtMail", bizPromoterVo.getPmtMail());
-
-			}
+				// 이메일 인증된 사용자의 경우
+				if(chkEmailYN){
+					log.info("프로모터 로그인 성공");
+					mv.setViewName("redirect:/Promoter/Index");
+					request.getSession().setAttribute("bizPromoter", bizPromoterVo);
+				}else{
+					// 이메일 인증이 안되어있는 사용자라면 인증 페이지로 redirect
+					log.info("이메일 인증되지 않은 사용자 로그인");
+					mv.setViewName("redirect:/Promoter/Facilitator/PmtCertifyPage.mwav");
+					rtr.addFlashAttribute("msg", "이메일 인증이 필요합니다.");
+					rtr.addFlashAttribute("promoter_id", bizPromoterVo.getPromoter_id());
+					rtr.addFlashAttribute("pmtMail", bizPromoterVo.getPmtMail());
+				}
+				break;
+			case 2:
+				log.info("프로모터 로그인 실패(아이디, 비밀번호 오류)");
+				mv.setViewName("redirect:/Promoter/Facilitator/PmtLogin.mwav");
+				rtr.addFlashAttribute("msg", "비밀번호와 아이디를 확인해주세요");
+				break;
+			case 3:
+				log.info("프로모터 로그인 실패(리캡챠 유효성 검증 실패)");
+				mv.setViewName("redirect:/Promoter/Facilitator/PmtLogin.mwav");
+				rtr.addFlashAttribute("msg", "로봇으로 감지되었습니다. 다시 시도해주세요");
+				break;
 
 		}
 		return mv;
@@ -240,74 +227,5 @@ public class PmtFacilitatorController {
 			log.info("세션에 프로모터 로그인 정보가 없어 로그아웃하지 못하였습니다");
 		}
 		return Status.OK;
-	}
-
-	public boolean recaptcha(String token){
-		//TODO: 임시용으로 VerifyRecaptcha 의 메소드 기능만 빼와서 사용. 추후 교체요망
-		final String SECRET_KEY = "6LdhTbYbAAAAAHFtrOEfPcyjW7XwgWGwxY0RrLVe";
-		final String SITE_VERIFY_URL = "https://www.google.com/recaptcha/api/siteverify";
-
-		if (token == null || token.length() == 0) {
-			return false;
-		}
-
-		try {
-			URL verifyUrl = new URL(SITE_VERIFY_URL);
-
-			// Open Connection to URL
-			HttpsURLConnection conn = (HttpsURLConnection) verifyUrl
-					.openConnection();
-
-			// Add Request Header
-			conn.setRequestMethod("POST");
-			//conn.setRequestProperty("User-Agent", "Mozilla/5.0");
-			//conn.setRequestProperty("Accept-Language", "en-US,en;q=0.5");
-
-			// Data will be sent to the server.
-			String postParams = "secret=" + SECRET_KEY + "&response="
-					+ token;
-
-			// Send Request
-			conn.setDoOutput(true);
-
-			// Get the output stream of Connection
-			// Write data in this stream, which means to send data to Server.
-			OutputStream outStream = conn.getOutputStream();
-			outStream.write(postParams.getBytes());
-
-			outStream.flush();
-			outStream.close();
-
-			// Response code return from server.
-			// HTTP.STATUS CODE
-			int responseCode = conn.getResponseCode();
-			System.out.println("responseCode=" + responseCode);
-
-			// Get the InputStream from Connection to read data sent from the
-			// server.
-			InputStream is = conn.getInputStream();
-
-			JsonReader jsonReader = Json.createReader(is);
-			JsonObject jsonObject = jsonReader.readObject();
-			jsonReader.close();
-
-			// ==> {"success": true}
-			//예)
-			/*{
-				  "success": true|false,
-				  "challenge_ts": timestamp,  // timestamp of the challenge load (ISO format yyyy-MM-dd'T'HH:mm:ssZZ)
-				  "hostname": string,         // the hostname of the site where the reCAPTCHA was solved
-				  "error-codes": [...]        // optional
-				}*/
-
-			System.out.println("Response: " + jsonObject);
-
-			boolean success = jsonObject.getBoolean("success");
-			return success;
-		} catch (Exception e) {
-			e.printStackTrace();
-			return false;
-		}
-
 	}
 }
